Implements ast navigation
In child/prev/next(), should we limit to the selection given in new()?
This selects around, what about inside?
What if the user only wants to move the head?
We could also do this with tree sitter queries, making it harder for the
user but enabling more patterns (see Helix).
That would be nice to have separators as well, so `fn add(a: u8, b: u8)`
have children, but how?